+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.10.1
 
 * fix another encoding issue, fixes again [#327](https://github.com/cars10/elasticvue/issues/327)
+* fix: redirect after adding first cluster
 
 ## 1.10.0